Finding cycle and removingthem, incomplete
Implementation of the algorithm that finds cylcles in the graph and removes the weight of the cycle from the graph and removes edges with weight 0. But there is a bug in the way the edges are recorded.

+++ flow_decomposition/src/edge.rs
@@ -2,7 +2,7 @@
 
 
 pub struct Edge {
-    pub id: i64,
+    pub id: usize,
     pub start_node: usize,
     pub end_node: usize,
     pub weight: i64,
